Fix error with search reducer and update reducer/index.js
