+++ src/reducers/index.js
@@ -1,8 +1,6 @@
-export const queriedBooks = (state = [], action) => {
-  switch (action.type) {
-    case 'SEARCH_BOOKS':
-      return [...action.foundBooks]
-    default:
-      return state;
-  }
-}
+import { combineReducers } from 'redux';
+import { queriedBooks } from './search';
+
+export const rootReducer = combineReducer({
+  queriedBooks
+});
